+++ lib/langchain/assistant/llm/adapters/openai.rb
@@ -1,8 +1,17 @@
+# frozen_string_literal: true
+
 module Langchain
   class Assistant
     module LLM
       module Adapters
-        class MistralAI < Base
+        class OpenAI < Base
+          # Build the chat parameters for the OpenAI LLM
+          #
+          # @param tools [Array] The tools to use
+          # @param instructions [String] The system instructions
+          # @param messages [Array] The messages
+          # @param tool_choice [String] The tool choice
+          # @return [Hash] The chat parameters
           def build_chat_params(tools:, instructions:, messages:, tool_choice:)
             params = {messages: messages}
             if tools.any?
@@ -12,8 +21,16 @@ module Langchain
             params
           end
 
+          # Build a OpenAI message
+          #
+          # @param role [String] The role of the message
+          # @param content [String] The content of the message
+          # @param image_url [String] The image URL
+          # @param tool_calls [Array] The tool calls
+          # @param tool_call_id [String] The tool call ID
+          # @return [Messages::OpenAIMessage] The OpenAI message
           def build_message(role:, content: nil, image_url: nil, tool_calls: [], tool_call_id: nil)
-            Langchain::Messages::MistralAIMessage.new(role: role, content: content, image_url: image_url, tool_calls: tool_calls, tool_call_id: tool_call_id)
+            Messages::OpenAIMessage.new(role: role, content: content, image_url: image_url, tool_calls: tool_calls, tool_call_id: tool_call_id)
           end
 
           # Extract the tool call information from the OpenAI tool call hash
@@ -36,14 +53,17 @@ module Langchain
             [tool_call_id, tool_name, method_name, tool_arguments]
           end
 
+          # Build the tools for the OpenAI LLM
           def build_tools(tools)
             tools.map { |tool| tool.class.function_schemas.to_openai_format }.flatten
           end
 
+          # Get the allowed assistant.tool_choice values for OpenAI
           def allowed_tool_choices
             ["auto", "none"]
           end
 
+          # Get the available tool names for OpenAI
           def available_tool_names(tools)
             build_tools(tools).map { |tool| tool.dig(:function, :name) }
           end